Update patch set 1
Patch Set 1:

Hey Daniel just checking on this. Are you waiting to merge this until Frankfurt is cut? I don't think there's a lot of urgency on it, so that's fine. I'm just planning to back port to el alto once its merged, so if frankfurt is cut first ill back port there as well.

Patch-set: 1