Update patch set 2
Patch Set 2: -Code-Review

> Patch Set 2:
> 
> Would prefere that the subject of the patch commit matches what is done.
> Seems that Python2.x tests are removed, which is not in title. 
> Difficult to search with log commits oneliners

The reason for this patch was not to remove the Python 2.x tests, but if you upgrade the robot framework versions to those needed to work reliably under Python3, the version 2.x tests will fail, and you will have a circular dependency with the verify jobs (testing code with Python 2.x) preventing subsequent patches from being merged. Removing the Python 2.x tests was not the primary reason for this patch. And it is mentioned on the next line down. Don't really understand the reason for the complaint here.

Patch-set: 2
Reviewer: Gerrit User 8049 <8049@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0